Azure_Functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
psycopg[binary]==3.2.6
SQLAlchemy==2.0.40
pydantic==2.11.3
